+++ packages/@react-stately/calendar/src/index.ts
@@ -10,6 +10,9 @@
  * governing permissions and limitations under the License.
  */
 
-export * from './types';
-export * from './useCalendarState';
-export * from './useRangeCalendarState';
+export {useCalendarState} from './useCalendarState';
+export {useRangeCalendarState} from './useRangeCalendarState';
+
+export type {CalendarStateOptions} from './useCalendarState';
+export type {RangeCalendarStateOptions} from './useRangeCalendarState';
+export type {CalendarState, RangeCalendarState} from './types';
